+++ packages/apptive_grid_core/lib/src/model/data_type.dart
@@ -51,8 +51,11 @@ enum DataType {
   /// Type for an email
   email(backendName: 'email'),
 
-  /// Type for a phone number
-  phoneNumber(backendName: 'phoneNumber');
+  /// Type for a phonenumber
+  phoneNumber(backendName: 'phoneNumber'),
+
+  /// Type for a phonenumber
+  signature(backendName: 'signature');
 
   /// Define a Datatype with a corresponding [backendName]
   const DataType({