+++ packages/apptive_grid_core/lib/src/network/attachment_processor.dart
@@ -77,7 +77,7 @@ class AttachmentProcessor {
     Uri? smallThumbnail;
     Uri? largeThumbnail;
 
-    if (type.startsWith('image/')) {
+    if (img.getDecoderForNamedImage(name) != null) {
       smallThumbnail = _generateUri(config);
       largeThumbnail = _generateUri(config);
     }
@@ -180,28 +180,18 @@ class AttachmentProcessor {
           ],
         );
         return uploads.first;
-      } else {
-        return _uploadFile(
-          baseUri: baseUploadUri,
-          headers: uploadHeaders,
-          createBytes: () async =>
-              attachmentAction.byteData ??
-              await File(attachmentAction.path!).readAsBytes(),
-          name: attachmentAction.attachment.url.pathSegments.last,
-          type: attachmentAction.attachment.type,
-        );
       }
-    } else {
-      return _uploadFile(
-        baseUri: baseUploadUri,
-        headers: uploadHeaders,
-        createBytes: () async =>
-            attachmentAction.byteData ??
-            await File(attachmentAction.path!).readAsBytes(),
-        name: attachmentAction.attachment.url.pathSegments.last,
-        type: attachmentAction.attachment.type,
-      );
     }
+
+    return _uploadFile(
+      baseUri: baseUploadUri,
+      headers: uploadHeaders,
+      createBytes: () async =>
+          attachmentAction.byteData ??
+          await File(attachmentAction.path!).readAsBytes(),
+      name: attachmentAction.attachment.url.pathSegments.last,
+      type: attachmentAction.attachment.type,
+    );
   }
 
   Future<http.Response> _uploadFile({