+++ packages/apptive_grid_form/lib/src/widgets/apptive_grid_form_widgets.dart
@@ -17,6 +17,7 @@ export 'package:apptive_grid_form/src/widgets/form_widget/geolocation_form_widge
 export 'package:apptive_grid_form/src/widgets/form_widget/integer_form_widget.dart';
 export 'package:apptive_grid_form/src/widgets/form_widget/multi_cross_reference_form_widget.dart';
 export 'package:apptive_grid_form/src/widgets/form_widget/phone_number_form_widget.dart';
+export 'package:apptive_grid_form/src/widgets/form_widget/signature_form_widget.dart';
 export 'package:apptive_grid_form/src/widgets/form_widget/text_form_widget.dart';
 export 'package:apptive_grid_form/src/widgets/form_widget/uri_form_widget.dart';
 export 'package:apptive_grid_form/src/widgets/form_widget/user_form_widget.dart';