+++ packages/apptive_grid_form/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.13.2
-* Add support for `DataType.email`, `DataType.phoneNumber`
+* Add support for `DataType.email`, `DataType.phoneNumber` and `DataType.signature`
 
 * ## 0.13.1+1
 * Update `image_picker`